+++ src/core/analysis-context.ts
@@ -14,6 +14,7 @@
 
 import * as path from 'path';
 
+
 import {ForkOptions, LazyEdgeMap, NoKnownParserError, Options, ScannerTable} from '../core/analyzer';
 import {CssCustomPropertyScanner} from '../css/css-custom-property-scanner';
 import {CssParser} from '../css/css-parser';
@@ -169,10 +170,10 @@ export class AnalysisContext {
    * Internal analysis method called when we know we need to fork.
    */
   private async _analyze(resolvedUrls: string[]): Promise<AnalysisContext> {
-    const analysisComplete = (async() => {
+    const analysisComplete = (async () => {
       // 1. Load and scan all root documents
       const scannedDocumentsOrWarnings =
-          await Promise.all(resolvedUrls.map(async(url) => {
+          await Promise.all(resolvedUrls.map(async (url) => {
             try {
               const scannedResult = await this.scan(url);
               this._cache.failedDocuments.delete(url);
@@ -186,7 +187,7 @@ export class AnalysisContext {
             }
           }));
       const scannedDocuments = scannedDocumentsOrWarnings.filter(
-          (d) => d != null) as ScannedDocument[];
+                                   (d) => d != null) as ScannedDocument[];
       // 2. Run per-document resolution
       const documents = scannedDocuments.map((d) => this.getDocument(d.url));
       // TODO(justinfagnani): instead of the above steps, do:
@@ -244,7 +245,7 @@ export class AnalysisContext {
     const document = new Document(scannedDocument, this, analysisResult);
     this._cache.analyzedDocuments.set(resolvedUrl, document);
     this._cache.analyzedDocumentPromises.getOrCompute(
-        resolvedUrl, async() => document);
+        resolvedUrl, async () => document);
 
     document.resolve();
     return document;
@@ -308,13 +309,14 @@ export class AnalysisContext {
    */
   private async _scanLocal(resolvedUrl: string): Promise<ScannedDocument> {
     return this._cache.scannedDocumentPromises.getOrCompute(
-        resolvedUrl, async() => {
+        resolvedUrl, async () => {
           try {
             const parsedDoc = await this._parse(resolvedUrl);
             const scannedDocument = await this._scanDocument(parsedDoc);
 
-            const imports = scannedDocument.getNestedFeatures().filter(
-                (e) => e instanceof ScannedImport) as ScannedImport[];
+            const imports =
+                scannedDocument.getNestedFeatures().filter(
+                    (e) => e instanceof ScannedImport) as ScannedImport[];
 
             // Update dependency graph
             const importUrls = imports.map((i) => this.resolveUrl(i.url));
@@ -333,10 +335,11 @@ export class AnalysisContext {
    */
   async scan(resolvedUrl: string): Promise<ScannedDocument> {
     return this._cache.dependenciesScannedPromises.getOrCompute(
-        resolvedUrl, async() => {
+        resolvedUrl, async () => {
           const scannedDocument = await this._scanLocal(resolvedUrl);
-          const imports = scannedDocument.getNestedFeatures().filter(
-              (e) => e instanceof ScannedImport) as ScannedImport[];
+          const imports =
+              scannedDocument.getNestedFeatures().filter(
+                  (e) => e instanceof ScannedImport) as ScannedImport[];
 
           // Scan imports
           for (const scannedImport of imports) {
@@ -451,7 +454,7 @@ export class AnalysisContext {
    */
   private async _parse(resolvedUrl: string): Promise<ParsedDocument> {
     return this._cache.parsedDocumentPromises.getOrCompute(
-        resolvedUrl, async() => {
+        resolvedUrl, async () => {
           const content = await this.load(resolvedUrl);
           const extension = path.extname(resolvedUrl).substring(1);
           return this._parseContents(extension, content, resolvedUrl);