+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Scan for CSS custom variable uses and assignments.
 * Fix value of reflectToAttribute for polymer properties
 * Remove scriptElement from PolymerElement
+* Introduce getDocumentContaining to find containing inline document for a feature
 
 ## [2.2.2] - 2017-07-20